[REFACTOR/#227] ApiCallUtils, ApiError을 core:network 모듈로 이전
