+++ core/network/src/main/java/com/yapp/network/model/ApiError.kt
@@ -1,4 +1,4 @@
-package com.yapp.data.remote.utils
+package com.yapp.network.model
 
 data class ApiError(
     override val message: String,
