+++ data/src/main/java/com/yapp/data/remote/datasource/UserInfoDataSourceImpl.kt
@@ -3,7 +3,7 @@ package com.yapp.data.remote.datasource
 import com.yapp.data.remote.dto.request.UpdateUserInfoRequest
 import com.yapp.data.remote.dto.response.UserResponse
 import com.yapp.data.remote.service.ApiService
-import com.yapp.data.remote.utils.safeApiCall
+import com.yapp.network.utils.safeApiCall
 import javax.inject.Inject
 
 class UserInfoDataSourceImpl @Inject constructor(
